requirements.txt
python-dotenv
flask
Faker
SQLAlchemy
flask_sqlalchemy
